+++ src/main/java/it/tredi/msa/mailboxmanager/MailboxManager.java
@@ -18,17 +18,35 @@ import it.tredi.msa.entity.AuditMailboxRunStatus;
 import it.tredi.msa.repository.AuditMailboxRunRepository;
 
 /**
- * Thread di gestione di una casella di posta elettronica. Classe astratta che si occupa
+ * Thread di gestione di una casella di posta elettronica. Classe astratta che si occupa di leggere il contenuto delle
+ * mailbox e processare i singoli messaggi individuati al loro interno
  */
 public abstract class MailboxManager implements Runnable {
 	
-	private MailboxConfiguration configuration;
-	private MailReader mailReader;
-	private boolean shutdown = false;
 	private static final Logger logger = LogManager.getLogger(MailboxManager.class.getName());
+	
+	/**
+	 * Configurazione della mailbox
+	 */
+	private MailboxConfiguration configuration;
+	
+	private MailReader mailReader;
+	
+	/**
+	 * Identifica se e' stato richiesto lo spegnimento del servizio
+	 */
+	private boolean shutdown = false;
+	
 	private AuditMailboxRun auditMailboxRun;
+	
+	/**
+	 * Identifica se l'attivita' di elaborazione di una mailbox e' al momento in corso
+	 */
 	private boolean running;
 	
+	/**
+	 * Repository di audit delle mailbox
+	 */
 	private AuditMailboxRunRepository auditMailboxRunRepository;
 	
 	private final static int MAILREADER_CONNECTION_ATTEMPTS = 3;
@@ -60,6 +78,9 @@ public abstract class MailboxManager implements Runnable {
 		this.running = running;
 	}
 
+	/**
+	 * Inizializzazione del manager
+	 */
 	public void init() {
 		this.auditMailboxRunRepository = ContextProvider.getBean(AuditMailboxRunRepository.class);
 		this.customInit();
@@ -69,7 +90,10 @@ public abstract class MailboxManager implements Runnable {
 	 * Eventuale inizializzazione custom dell'oggetto manager
 	 */
 	public abstract void customInit(); 
-		
+	
+	/**
+	 * Esecuzione dell'elaborazione della mailbox
+	 */
 	@Override
     public void run() {
 		running = true;
@@ -98,6 +122,9 @@ public abstract class MailboxManager implements Runnable {
     	}
     }
     
+	/**
+	 * Chiusura del processo di elaborazione della mailbox
+	 */
     public void shutdown() {
     	try {
         	shutdown = true;
@@ -118,6 +145,9 @@ public abstract class MailboxManager implements Runnable {
     	}
     }	
     
+    /**
+     * Elaborazione della mailbox (lettura messaggi e conversione in documenti)
+     */
     public void processMailbox() {
     	try {
         	if (logger.isDebugEnabled())
@@ -194,6 +224,10 @@ public abstract class MailboxManager implements Runnable {
     	}
     }
     
+    /**
+     * Apertura della sessione di lavoro
+     * @throws Exception
+     */
     public void openSession() throws Exception {
     	if (logger.isDebugEnabled())
     		logger.debug("[" + configuration.getName() + "] opening mailReader connection");
@@ -220,6 +254,9 @@ public abstract class MailboxManager implements Runnable {
     		logger.debug("[" + configuration.getName() + "] mailReader connection opened");		
     }
 
+    /**
+     * Chiusura della sessione di lavoro
+     */
     public void closeSession() {
     	//audit - mailbox run
     	if (auditMailboxRun != null) { //call it just one time
@@ -246,10 +283,22 @@ public abstract class MailboxManager implements Runnable {
 		}
     }    
     
+    /**
+     * Processa il messaggio email letto dalla casella e produce un oggetto ParsedMessage (analisi di tutte le 
+     * parti del messaggio)
+     * @param message
+     * @return
+     * @throws Exception
+     */
     public ParsedMessage parseMessage(Message message) throws Exception {
     	return new ParsedMessage(message);
     }
     
+    /**
+     * Elaborazione del messaggio parsato (conversione in documento e salvataggio)
+     * @param parsedMessage
+     * @throws Exception
+     */
     public void processMessage(ParsedMessage parsedMessage) throws Exception {
     	if (logger.isDebugEnabled())
     		logger.debug("[" + configuration.getName() + "] processMessage() called");
@@ -278,6 +327,11 @@ public abstract class MailboxManager implements Runnable {
     		logger.debug("[" + configuration.getName() + "] processMessage() done");
     }
 
+    /**
+     * Gestione di un errore riscontrato sull'elaborazione di un messaggio (scrittura dell'errore su audit, notifica, ecc.)
+     * @param t
+     * @param obj
+     */
     public void handleError(Throwable t, Object obj) {
     	if (shutdown)
     		logger.warn("[" + configuration.getName() + "] exception during shutdown... ignoring error", t);